Moved ParametricModel repr unit tests to a separate file
